warn user on overwrite; prefill selection

+++ src/store.ts
@@ -5,7 +5,11 @@ import * as fs from 'fs';
 import * as path from 'path';
 
 function getStoreFile() {
-    var config: string = vscode.workspace.getConfiguration().get('rootPath');
+    // TODO: created by admin @ 2017-10-19 11:07:33
+    // need to add this to the config
+    // get from config 
+    var config: string = vscode.workspace.getConfiguration().get('store.rootPath');
+    // 
     if (!config) {
         config = process.env[(process.platform == 'win32') ? 'USERPROFILE' : 'HOME'];
     }
@@ -19,7 +23,7 @@ function getStoreFile() {
 export const activate = (context: vscode.ExtensionContext) => {
 
     const newStore = vscode.commands.registerCommand('store.new', () => {
-
+        const editor = vscode.window.activeTextEditor;
         vscode.window.showInputBox({
             placeHolder: "Enter new key to store",
             ignoreFocusOut: true
@@ -27,21 +31,26 @@ export const activate = (context: vscode.ExtensionContext) => {
             if (!key) {
                 return
             }
-            // TODO: created by admin @ 2017-10-16 19:08:57
-            // if the key already exists user has to be warned. 
+            const storeFilePath = getStoreFile()
+            const storeData = JSON.parse(fs.readFileSync(storeFilePath, 'utf8'));
+
+            // if the key already exists user has to be warned.
+            if( new Set(Object.keys(storeData) ).has(key) ){
+                vscode.window.showWarningMessage('Key already exists. Try removing it and add again.')
+                return;
+            }
             vscode.window.showInputBox({
                 ignoreFocusOut: true,
-                placeHolder: "Enter value to store"
+                placeHolder: "Enter value to store",
+                // prefill the selected text
+                value: editor ? editor.document.getText(editor.selection) : '',
+                prompt: "Enter value to store"
             }).then(value => {
                 if (!value) {
                     return
                 }
-                console.log(key, value)
                 //  Store the data in a json file
-                const storeFilePath = getStoreFile()
-                const storeData = JSON.parse(fs.readFileSync(storeFilePath, 'utf8'));
                 storeData[key] = value;
-                console.log('storeData', storeData)
                 fs.writeFileSync(storeFilePath, JSON.stringify(storeData), 'utf8');
             })
         });
@@ -82,7 +91,7 @@ export const activate = (context: vscode.ExtensionContext) => {
                 delete storeData[key]
                 // Write the file to disk
                 fs.writeFileSync( storeFilePath, JSON.stringify(storeData), 'utf8');
-                vscode.window.showInformationMessage(`Removed ${key} from store`);
+                vscode.window.showInformationMessage(`Removed '${key}' from store`);
             })
     })
